La clase Item ahora tiene ItemName como id

+++ Assets/Scripts/GameScene/HUD/SlotsController.cs
@@ -38,7 +38,7 @@ public class SlotsController : MonoBehaviour
                     }
                     else
                     {
-                        image.sprite = itemSprites[items[i].id]; // Cargar la imagen del item
+                        image.sprite = itemSprites[(int)items[i].id]; // Cargar la imagen del item
                         image.gameObject.SetActive(true);
 
                         GameObject text = image.gameObject.transform.parent.GetChild(image.gameObject.transform.parent.childCount - 1).gameObject;
